Represent STEP INTEGER as Long

+++ SceneJS/src/org/bimserver/geometry/jsonshell/SceneJsShellSerializer.java
@@ -558,7 +558,7 @@ public class SceneJsShellSerializer extends AbstractGeometrySerializer {
 		if (object.isSetRefLatitude()) {
 			jsonWriter.name("Ref Latitude");
 			jsonWriter.beginArray();
-			for (Integer val : object.getRefLatitude()) {
+			for (Long val : object.getRefLatitude()) {
 				jsonWriter.value(val.toString());
 			}
 			jsonWriter.endArray();
@@ -566,7 +566,7 @@ public class SceneJsShellSerializer extends AbstractGeometrySerializer {
 		if (object.isSetRefLongitude()) {
 			jsonWriter.name("Ref Longtitude");
 			jsonWriter.beginArray();
-			for (Integer val : object.getRefLongitude()) {
+			for (Long val : object.getRefLongitude()) {
 				jsonWriter.value(val.toString());
 			}
 			jsonWriter.endArray();
